hw4/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-logging
google-cloud-pubsub